apps/research-auditor/pyproject.toml
[project]
name = "research-auditor"
version = "0.1.0"
description = "Stable multi-agent research environment with industrial pipeline"
requires-python = ">=3.12,<3.13"

dependencies = [
  "requests>=2.31,<3",
  "python-dotenv>=1,<2",

  # Keep your deliberate pydantic ceiling if you know 2.12 breaks your stack:
  "pydantic>=2.11.9,<2.12",

  # pydantic-ai is major 1 now; cap at next major:
  "pydantic-ai-slim[openai]>=1.59,<2",


  "langgraph>=1.0.6,<1.1",

  "supabase>=2.28,<3",

  # If you’re okay moving with logfire’s current major:
  "logfire>=4.24,<5",

  "fastapi>=0.115,<1",
  "uvicorn[standard]>=0.32,<1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8",
  "mypy>=1.13",
  "bandit[toml]>=1.7",
  "pytest>=8",
  "pytest-cov>=4",
]

[tool.uv]
managed = true
package = false
